feat : MVC 분리
- output 로직을 모두 RacingGameView로 분리
- 그외 Main및 RacingGame의 로직이동
- RacingGameView에서 model과의 의존성 제거
